Add files via upload
Problem Statement:
Mashable is a global, multi – platform media and entertainment company. Powered by its own proprietary technology, Mashable is the go-to source for tech, digital culture and entertainment content for its dedicated and influential audience around the globe. Just like any other media company its success depends on the popularity of articles. And one of the key metrics to measure popularity is number of shares done on the article. Over the period of few years, Mashable has collected data on around 40, 000 articles. The goal is to do analysis and modelling to predict the popularity of an article given the input parameters.